requirements.txt
# Web Scraping and Browser Automation
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0

# Database
pymongo>=4.5.0

# DNS Resolution
dnspython>=2.4.0

# Excel Export
xlsxwriter>=3.1.0
openpyxl>=3.1.0

# HTML Parsing
lxml>=4.9.0

# URL Parsing (built-in urllib.parse)
# CSV handling (built-in csv)
# Threading (built-in threading)
# Regular Expressions (built-in re) 